requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
openpyxl>=3.1.2
numpy>=1.26.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
gspread>=5.10.0 